fix: rename tensorflow to keras
