+++ configs/ops/prediction/op_prediction_base.yaml
@@ -1,5 +1,5 @@
 model_loader:
-  _target_: niceml.dlframeworks.tensorflow.kerasmodelloader.KerasModelLoader
+  _target_: niceml.dlframeworks.keras.kerasmodelloader.KerasModelLoader
 prediction_steps: ${niceml.to_int:${oc.env:PREDICTION_STEPS,null}}
 prediction_function:
-  _target_: niceml.dlframeworks.tensorflow.predictionfunctions.keraspredictionfunction.KerasPredictionFunction
+  _target_: niceml.dlframeworks.keras.predictionfunctions.keraspredictionfunction.KerasPredictionFunction
