+++ configs/ops/train/op_train_cls_binary.yaml
@@ -25,12 +25,12 @@ data_validation:
     sub_dir: validation
   set_name: validation
 model:
-  _target_: niceml.dlframeworks.tensorflow.models.mobilenet.OwnMobileNetModel
+  _target_: niceml.dlframeworks.keras.models.mobilenet.OwnMobileNetModel
   final_activation: sigmoid
 learner:
-  _target_: niceml.dlframeworks.tensorflow.learners.defaultlearner.DefaultLearner
+  _target_: niceml.dlframeworks.keras.learners.keraslearner.KerasLearner
   model_compiler:
-    _target_: niceml.dlframeworks.tensorflow.modelcompiler.defaultmodelcompiler.DefaultModelCompiler
+    _target_: niceml.dlframeworks.keras.modelcompiler.defaultmodelcompiler.DefaultModelCompiler
     loss: binary_crossentropy
     # Optimizer used in the experiment
     optimizer:
