+++ niceml/dlframeworks/keras/models/unets.py
@@ -14,8 +14,8 @@ from niceml.data.datadescriptions.inputdatadescriptions import InputImageDataDes
 from niceml.data.datadescriptions.outputdatadescriptions import (
     OutputImageDataDescription,
 )
-from niceml.dlframeworks.tensorflow.models.layerfactory import LayerFactory
-from niceml.dlframeworks.tensorflow.models.premodellayers import add_premodel_layers
+from niceml.dlframeworks.keras.models.layerfactory import LayerFactory
+from niceml.dlframeworks.keras.models.premodellayers import add_premodel_layers
 from niceml.mlcomponents.models.modelfactory import ModelFactory
 from niceml.utilities.commonutils import check_instance
