+++ pyproject.toml
@@ -74,6 +74,7 @@ tensorflow-metal = {version = "~1.0", optional = true}
 requests = "^2.31.0"
 cryptography = "^41.0.0"
 dagster-mlflow = "^0.20.14"
+grpcio = "1.55.3"
 
 
 [tool.poetry.extras]
