+++ template/configs/ops/train/callbacks/callbacks_base.yaml
@@ -4,12 +4,12 @@
 
 _target_: niceml.mlcomponents.callbacks.callbackinitializer.CallbackInitializer
 callback_list:
-  - _target_: niceml.dlframeworks.tensorflow.callbacks.callback_factories.InitCallbackFactory
+  - _target_: niceml.dlframeworks.keras.callbacks.callback_factories.InitCallbackFactory
     callback:
-      _target_: niceml.dlframeworks.tensorflow.callbacks.nancheckcallback.LossNanCheckCallback
-  - _target_: niceml.dlframeworks.tensorflow.callbacks.callback_factories.LoggingOutputCallbackFactory
+      _target_: niceml.dlframeworks.keras.callbacks.nancheckcallback.LossNanCheckCallback
+  - _target_: niceml.dlframeworks.keras.callbacks.callback_factories.LoggingOutputCallbackFactory
 callback_dict:
   save_model:
     # Stores the model after each epoch
-    _target_: niceml.dlframeworks.tensorflow.callbacks.callback_factories.ModelCallbackFactory
+    _target_: niceml.dlframeworks.keras.callbacks.callback_factories.ModelCallbackFactory
     model_subfolder: models/model-id_{short_id}-ep{epoch:03d}.hdf5
